pyproject.toml
[project]
name = "blank-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=5.5.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "snowflake-connector-python>=3.16.0",
    "streamlit>=1.46.1",
]
